requirements.txt
dj-database-url==0.2.2
dj-static==0.0.5
django-decouple==2.1
django-filter==0.7
django-markdown==0.6.0
Django==1.6.1
djangorestframework==2.3.13
gunicorn==18.0
psycopg2==2.5.1
pytz==2013.9
South==0.8.4
static==0.4
Unipath==1.0
